(360 MSVC WIP) Use roundf instead of round - changed double to float in vce.cpp for function SetPixelFormat

+++ mednafen/pce/vce.cpp
@@ -381,9 +381,9 @@ void VCE::SetPixelFormat(const MDFN_PixelFormat &format)
   }
   else
   {
-   double y;
+   float y;
 
-   y = round(0.300 * r + 0.589 * g + 0.111 * b);
+   y = roundf(0.300 * r + 0.589 * g + 0.111 * b);
 
    if(y < 0)
     y = 0;
